- Moved code from utility to APITrackerService - Added statement for Create Index if not exists
