+++ src/utility/utility.ts
@@ -7,9 +7,6 @@ import { PermissionRequest } from "nodets-ms-core/lib/core/auth/model/permission
 import _ from "lodash";
 import { InputException } from "../exceptions/http/http-exceptions";
 import AdmZip from "adm-zip";
-import dbClient from "../database/data-source";
-import { APIUsageSummaryEntity, APIUsageDetailsEntity } from '../database/entity/api-usage-entity';
-import { Request } from 'express';
 
 export class Utility {
 
@@ -160,38 +157,5 @@ export class FileEntityStream extends Readable {
         fileStream.on('error', (err) => {
             this.emit('error', err);
         });
-    }
-}
-
-
-export class APIUsageUtility {
-    public static async saveAPIUsageSummary(req: Request): Promise<void> {
-        try {
-            const matchedRoute = req.route?.path || req.originalUrl;
-            await dbClient.query(APIUsageSummaryEntity.getUpsertApiUsageSummaryQuery(matchedRoute));
-        } catch (error) {
-            console.error(error);
-        }
-    }
-
-    public static async saveAPIUsageDetails(req: Request): Promise<void> {
-        try {
-            const matchedRoute = req.route?.path || req.originalUrl;
-            const request_params = {
-              ...req.body,
-              ...req.params,
-              ...req.query
-            }
-            const details = new APIUsageDetailsEntity({
-                endpoint: matchedRoute,
-                method: req.method,
-                client_ip: req.ip,
-                user_id: req.body.user_id,
-                request_params: request_params || null
-            })
-            await dbClient.query(APIUsageDetailsEntity.getCreateAPIUsageDetailsQuery(details));
-        } catch (error) {
-            console.error(error);
-        }
     }
 }
